pyproject.toml
[project]
name = "coding-env"
version = "0.1.0"
description = "Coding environment with bash, editor, and VNC tools for solving programming tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hud-python[agents]>=0.5.17",
    "click>=8.0.0",
    "mcp[cli]>=1.10.1",
    "numpy>=1.26.0",
    "packaging>=21.0",
    "pillow",
    "pydantic>=2.11.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff==0.11.2", "pre-commit>=3.5.0"]

[project.scripts]
hud_eval = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "I",  # isort
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
